# benchmarks/wagon: use wagon-0.5 and AOT mode