--- a/benchmarks/wagon/go.mod
+++ b/benchmarks/wagon/go.mod
@@ -2,4 +2,4 @@ module github.com/wasmerio/go-ext-wasm/benchmarks-wagon
 
 go 1.12
 
-require github.com/go-interpreter/wagon v0.4.0
+require github.com/go-interpreter/wagon v0.5.0