```
  $> benchstat wagon-0.4.txt wagon-0.5-new.txt
  name                          old time/op  new time/op  delta
  _Wagon_Nbody-8                 1.52s ± 1%   1.06s ± 1%  -30.15%  (p=0.000 n=8+10)
  _Wagon_Fibonacci_Recursive-8   2.84s ± 4%   1.71s ± 1%  -39.92%  (p=0.000 n=9+10)
  _Wagon_Pollard_Rho_128-8       1.91s ± 6%   0.94s ± 1%  -50.85%  (p=0.000 n=10+9)
```