@sbinet 2019-06-07
thanks to the awesome work of @twitchyliquid64, `wagon-0.5` is 30-50% faster when AOT mode is enabled. (just thought it would be interesting to bring that up in the `wagon` benchmarks :))

`wasmer` is faster, though :)
```
$> benchstat  wagon-0.5-new.txt wasmer.txt 
name                          old time/op  new time/op  delta
_Nbody-8                 1.05s ± 1%   0.03s ± 1%  -96.69%  (p=0.000 n=9+10)
_Fibonacci_Recursive-8   1.72s ± 1%   0.02s ± 1%  -98.63%  (p=0.000 n=9+9)
_Pollard_Rho_128-8       851ms ± 2%    30ms ± 1%  -96.47%  (p=0.000 n=10+9)
```