@Hywan 2019-06-07
I don't get the 30-50% you observe. What platform are you using? I'm on macOS.