@sbinet 2019-06-07
Yes. For the moment, only Linux amd64 is supported for the AOT mode.