@twitchyliquid64 2019-06-07
The restriction to Linux is mostly artificial. I need a way to map executable memory, which varies by platform.

I am using the mmap package though, which has support for Darwin & iirc Linux. I _think_ (untested) it should be fine to remove the platform check.