@twitchyliquid64 2019-06-07
If you try changing this (https://github.com/go-interpreter/wagon/blob/83c88855b793033e6bb0bb77bf9f3727a5dfeda5/exec/native_compile_nogae.go#L16) to Darwin, i think it will work.

As for why it's only 50% faster (rather than 5x) , I think that's because your benchmarks are comparatively branch-heavy, but wagons native compiler explicitly does not compile branches. I'm also missing support for a good number of instructions (including some basic ones like load memory, Ive been putting off implementing as I need to mitigate spectre).

The fact that cranelift gets wasmer to ~50x faster though shows Ive still got lots of headroom to improve performance. Perhaps I should go ahead and implement that peephole optimiser after all.