@Hywan 2019-06-11
I look forward to try the macOS/darwin support :-).