@syrusakbary 2021-02-05
Closing the PR as now seems stale. Happy to merge new PRs once the last feedback (macOS/darwin support) is addressed :)